Add method to check number of open files on unix-based systems
