+++ server/src/util/macros.rs
@@ -1,5 +1,5 @@
 /*
- * Created on Fri Jun 25 2021
+ * Created on Sat Jan 29 2022
  *
  * This file is a part of Skytable
  * Skytable (formerly known as TerrabaseDB or Skybase) is a free and open-source
@@ -7,7 +7,7 @@
  * vision to provide flexibility in data modelling without compromising
  * on performance, queryability or scalability.
  *
- * Copyright (c) 2021, Sayan Nandan <ohsayan@outlook.com>
+ * Copyright (c) 2022, Sayan Nandan <ohsayan@outlook.com>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published by
@@ -24,45 +24,11 @@
  *
 */
 
-/// # Unsafe unwrapping
-///
-/// This trait provides a method `unsafe_unwrap` that is potentially unsafe and has
-/// the ability to **violate multiple safety gurantees** that rust provides. So,
-/// if you get `SIGILL`s or `SIGSEGV`s, by using this trait, blame yourself.
-pub unsafe trait Unwrappable<T> {
-    /// Unwrap a _nullable_ (almost) type to get its value while asserting that the value
-    /// cannot ever be null
-    ///
-    /// ## Safety
-    /// The trait is unsafe, and so is this function. You can wreck potential havoc if you
-    /// use this heedlessly
-    ///
-    unsafe fn unsafe_unwrap(self) -> T;
-}
-
 #[macro_export]
 macro_rules! impossible {
     () => {
         core::hint::unreachable_unchecked()
     };
-}
-
-unsafe impl<T, E> Unwrappable<T> for Result<T, E> {
-    unsafe fn unsafe_unwrap(self) -> T {
-        match self {
-            Ok(t) => t,
-            Err(_) => impossible!(),
-        }
-    }
-}
-
-unsafe impl<T> Unwrappable<T> for Option<T> {
-    unsafe fn unsafe_unwrap(self) -> T {
-        match self {
-            Some(t) => t,
-            None => impossible!(),
-        }
-    }
 }
 
 #[macro_export]
@@ -193,51 +159,6 @@ macro_rules! afn_action {
             }
         )*
     };
-}
-
-pub mod compiler {
-    //! Dark compiler arts and hackery to defy the normal. Use at your own
-    //! risk
-
-    use core::mem;
-
-    #[cold]
-    #[inline(never)]
-    pub const fn cold() {}
-
-    pub const fn likely(b: bool) -> bool {
-        if !b {
-            cold()
-        }
-        b
-    }
-
-    pub const fn unlikely(b: bool) -> bool {
-        if b {
-            cold()
-        }
-        b
-    }
-
-    #[cold]
-    #[inline(never)]
-    pub const fn cold_err<T>(v: T) -> T {
-        v
-    }
-    #[inline(always)]
-    pub const fn hot<T>(v: T) -> T {
-        if false {
-            cold()
-        }
-        v
-    }
-
-    pub unsafe fn extend_lifetime<'a, 'b, T>(inp: &'a T) -> &'b T {
-        mem::transmute(inp)
-    }
-    pub unsafe fn extend_lifetime_mut<'a, 'b, T>(inp: &'a mut T) -> &'b mut T {
-        mem::transmute(inp)
-    }
 }
 
 #[macro_export]
